refactor: 드롭다운 생성 로직 모듈화 (#74)
