+++ src/lib/index.ts
@@ -1,6 +1,17 @@
 import { inputContainerVariants } from "@/lib/inputContainerVariant";
-import { cn } from "@/lib/utils";
+import {
+  cn,
+  createCohortsDropdownOptions,
+  createCoursesDropdownOptions,
+} from "@/lib/utils";
 import api from "@/lib/axios";
 import { buttonVariants } from "@/lib/buttonVariant";
 
-export { inputContainerVariants, cn, api, buttonVariants };
+export {
+  inputContainerVariants,
+  cn,
+  api,
+  buttonVariants,
+  createCohortsDropdownOptions,
+  createCoursesDropdownOptions,
+};
