+++ src/lib/utils.ts
@@ -1,6 +1,45 @@
-import { type ClassValue, clsx } from "clsx"
-import { twMerge } from "tailwind-merge"
+import type { DropdownOption } from "@/types";
+import type { AvailableCourse } from "@/types/student-enrollment-types";
+import { type ClassValue, clsx } from "clsx";
+import { twMerge } from "tailwind-merge";
 
 export function cn(...inputs: ClassValue[]) {
-  return twMerge(clsx(inputs))
+  return twMerge(clsx(inputs));
+}
+
+export function createCoursesDropdownOptions(
+  availableCourses: AvailableCourse[]
+) {
+  const newCourses: DropdownOption[] = [];
+
+  availableCourses.forEach((availableCourse) => {
+    const courseDropdownOption: DropdownOption = {
+      label: availableCourse.course.name,
+      value: String(availableCourse.course.id),
+    };
+
+    newCourses.push(courseDropdownOption);
+  });
+
+  return newCourses;
+}
+
+export function createCohortsDropdownOptions(
+  availableCourses: AvailableCourse[],
+  selectedCourseId: string
+) {
+  const newCohorts: DropdownOption[] = [];
+
+  availableCourses.forEach((availableCourse) => {
+    if (selectedCourseId === String(availableCourse.course.id)) {
+      const cohortDropdownOption: DropdownOption = {
+        label: `${availableCourse.cohort.number}기`,
+        value: String(availableCourse.cohort.id),
+      };
+
+      newCohorts.push(cohortDropdownOption);
+    }
+  });
+
+  return newCohorts;
 }
